Struct massa_ledger_worker::ledger::FinalLedger

pub struct FinalLedger {
    pub(crate) config: LedgerConfig,
    pub(crate) sorted_ledger: LedgerDB,
}

Represents a final ledger associating addresses to their balances, bytecode and data. The final ledger is part of the final state which is attached to a final slot, can be bootstrapped and allows others to bootstrap. The ledger size can be very high: it can exceed 1 terabyte. To allow for storage on disk, the ledger uses trees and has O(log(N)) access, insertion and deletion complexity.

Fields
config: LedgerConfig

ledger configuration

sorted_ledger: LedgerDB

ledger tree, sorted by address

Implementations

impl FinalLedger

pub fn new(config: LedgerConfig, with_final_state: bool) -> Self

Initializes a new FinalLedger by reading its initial state from file.