massa_ledger_worker

Module ledger_db

Copyright (c) 2022 MASSA LABS info@massa.net Module to interact with the disk ledger

Structs

InitialLedgerLoader 🔒

LedgerDB

Disk ledger DB module

Enums

LedgerSubEntry

Ledger sub entry enum

Functions

db_batch_byte_size 🔒

delete_datastore_entries 🔒

end_prefix 🔒

For a given start prefix (inclusive), returns the correct end prefix (non-inclusive). This assumes the key bytes are ordered in lexicographical order. Since key length is not limited, for some case we return None because there is no bounded limit (every keys in the series [], [255], [255, 255] …).

initial_ledger_file_error 🔒