massa_ledger_worker::ledger_db

Struct LedgerDB

pub struct LedgerDB {
    db: ShareableMassaDBController,
    thread_count: u8,
    key_serializer_db: KeySerializer,
    key_deserializer_db: KeyDeserializer,
    amount_serializer: AmountSerializer,
    version_serializer: U64VarIntSerializer,
    version_deserializer: U64VarIntDeserializer,
    bytecode_serializer: BytecodeSerializer,
    amount_deserializer: AmountDeserializer,
    bytecode_deserializer: BytecodeDeserializer,
    max_datastore_value_length: u64,
    max_datastore_key_length: u8,
}

Disk ledger DB module

Contains a RocksDB DB instance

Fields
db: ShareableMassaDBControllerthread_count: u8key_serializer_db: KeySerializerkey_deserializer_db: KeyDeserializeramount_serializer: AmountSerializerversion_serializer: U64VarIntSerializerversion_deserializer: U64VarIntDeserializerbytecode_serializer: BytecodeSerializeramount_deserializer: AmountDeserializerbytecode_deserializer: BytecodeDeserializermax_datastore_value_length: u64max_datastore_key_length: u8
Implementations

impl LedgerDB

pub fn new( db: ShareableMassaDBController, thread_count: u8, max_datastore_key_length: u8, max_datastore_value_length: u64, max_bytecode_size: u64, ) -> Self

Create and initialize a new LedgerDB.

Arguments

path: path to the desired disk ledger db directory

pub fn load_initial_ledger( &mut self, initial_ledger: HashMap<Address, LedgerEntry>, ) -> Result<(), LedgerError>

Loads the initial disk ledger from an in-memory map.

Arguments

pub fn load_initial_ledger_from_path( &mut self, path: &Path, ) -> Result<(), LedgerError>

Loads the initial disk ledger from a JSON file, one entry at a time.

fn flush_initial_ledger_batch(&self, batch: DBBatch, slot: Option<Slot>)

fn ensure_ledger_is_empty(&self) -> Result<(), LedgerError>

pub fn apply_changes_to_batch( &self, changes: LedgerChanges, batch: &mut DBBatch, )

Allows applying LedgerChanges to the disk ledger

Arguments

changes: ledger changes to be applied

batch: the batch to apply the changes to

pub fn get_sub_entry( &self, addr: &Address, ty: LedgerSubEntry, ) -> Option<Vec<u8>>

Get the given sub-entry of a given address.

Arguments

addr: associated address

ty: type of the queried sub-entry

Returns

An Option of the sub-entry value as bytes

pub fn get_datastore_keys( &self, addr: &Address, prefix: &[u8], start_key: Bound<Vec<u8>>, end_key: Bound<Vec<u8>>, count: Option<u32>, ) -> Option<BTreeSet<Vec<u8>>>

Get every key of the datastore for a given address.

Returns

A BTreeSet of the datastore keys

pub fn reset(&self)

pub fn is_key_value_valid( &self, serialized_key: &[u8], serialized_value: &[u8], ) -> bool

Deserializes the key and value, useful after bootstrap

impl LedgerDB

fn put_entry( &self, addr: &Address, ledger_entry: LedgerEntry, batch: &mut DBBatch, )

Add every sub-entry individually for a given entry.

Arguments

addr: associated address

ledger_entry: complete entry to be added

batch: the given operation batch to update

fn update_entry( &self, addr: &Address, entry_update: LedgerEntryUpdate, batch: &mut DBBatch, )

Update the ledger entry of a given address.

If the address has no canonical entry yet (no VERSION marker), a default entry is materialized first via Self::put_entry, matching the contract documented on Self::apply_changes_to_batch and the LedgerChanges Vacant helpers (F84).

Arguments

entry_update: a descriptor of the entry updates to be applied

batch: the given operation batch to update

fn delete_entry(&self, addr: &Address, batch: &mut DBBatch)

Delete every sub-entry associated to the given address.

Arguments

batch: the given operation batch to update